Tareas/sistema-ventas/ms-compras/requirements.txt
fastapi[standard]==0.115.12
pydantic-settings==2.9.1
asyncpg==0.30.0
sqlalchemy==2.0.41
pyjwt==2.10.1
aio_pika==9.5.5